+++ src/Storing/store.cpp
@@ -10,10 +10,10 @@ bool Store::is_init(){ return false; }
 void Store::init(){}
 
 bool Store::exists(queue<string> seq){ return false; }
-void Store::create(queue<string> seq, int nb){}
+void Store::create(queue<string> seq, long long int nb){}
 
-int Store::get_hits(queue<string> seq){ return 0; }
-void Store::add_hits(queue<string> seq, int nb){}
+long long int Store::get_hits(queue<string> seq){ return 0; }
+void Store::add_hits(queue<string> seq, long long int nb){}
 
 double Store::get_proba(queue<string> seq){ return 0.0; }
 string Store::next(queue<string> seq){ return ""; }
